Store MS data objects using the alabaster framework

The alabaster framework provides the methodology to save R objects to on-disk representations/ storage modes which are programming language independent (in contrast to e.g. R's RDS files). By using standard file formats such as JSON and HDF5, alabaster ensures that the data can also be read and imported by other programming languages such as Python or Javascript. This improves interoperability between application ecosystems.

The alabaster package defines the alabaster.base::saveObject() and alabaster.base::readObject() methods that have to be implemented for specific data classes to enable saving to or reading from alabaster-based file formats.

In addition, the MsStash package defines the AlabasterParam which can be used to write or read MS objects using the saveMsObject() and readMsObject() methods in alabaster format. This allows additional configurations and customizations to the export or import process. It is thus for example possible to specify the path to the original MS data files for on-disk MS representations such as the Spectra::MsBackendMzR which enables to import a stored object even if either the object or the original MS data files have been moved to a different directory or file system.

In order to enable alabaster-based import/export, the respective methods have to be implemented for the class. See the package vignette for examples and details.

Usage

AlabasterParam(path = tempdir())

Arguments

path

character(1) with the name of the directory where the MS data object should be saved to or from which it should be restored. Importantly, path should point to a new folder, i.e. a directory that does not already exist.

Value

For AlabasterParam(): an instance of AlabasterParam class. For readObject() the exported object in the specified path (depending on the type of object defined in the OBJECT file in the path. For readMsObject() the exported data object, defined with the function's first parameter, from the specified path. saveObject() and saveMsObject() don't return anything.

Details

Importantly, it is only possible to save one object in one directory. To overwrite an existing stored object in a folder, that folder has to be deleted beforehand.

See also

Author

Johannes Rainer, Philippine Louail

Examples

## Create an AlabasterParam object
a <- AlabasterParam(path = tempdir())
a
#> Object of class:  AlabasterParam 
#>  Parameters:
#>  - path: [1] "/tmp/RtmpzsP6Ka"

## See the package vignette for example implementation and usage.